getting duration with ffprobe

+++ internal/processor/mock/processor_mock.go
@@ -41,10 +41,10 @@ func (m *MockVideoProcessor) EXPECT() *MockVideoProcessorMockRecorder {
 }
 
 // GetDuration mocks base method.
-func (m *MockVideoProcessor) GetDuration(ctx context.Context, inputPath string) (int64, error) {
+func (m *MockVideoProcessor) GetDuration(ctx context.Context, inputPath string) (float64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetDuration", ctx, inputPath)
-	ret0, _ := ret[0].(int64)
+	ret0, _ := ret[0].(float64)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
